chore: updates to utils for apex debug
