+++ packages/salesforcedx-utils/src/cli/index.ts
@@ -1,3 +1,7 @@
+export { CliCommandExecution } from './cliCommandExecution';
+export { CliCommandExecutor } from './cliCommandExecutor';
 export { CommandBuilder } from './commandBuilder';
 export { CommandOutput } from './commandOutput';
 export { ForceOrgDisplay } from './forceOrgDisplay';
+export { SfdxCommandBuilder } from './sfdxCommandBuilder';
+export { ForceConfigGet } from './forceConfigGet';
