+++ packages/salesforcedx-utils/src/constants/constants.ts
@@ -13,3 +13,5 @@ export const BASE_FILE_NAME = 'i18n';
 export const BASE_FILE_EXTENSION = 'js';
 export const DEFAULT_LOCALE = 'en';
 export const MISSING_LABEL_MSG = '!!! MISSING LABEL !!!';
+export const SFDX_CONFIG_ISV_DEBUGGER_SID = 'isvDebuggerSid';
+export const SFDX_CONFIG_ISV_DEBUGGER_URL = 'isvDebuggerUrl';
