+++ packages/salesforcedx-utils/src/helpers/extractJsonObject.ts
@@ -1,0 +1,5 @@
+export function extractJsonObject(str: string): any {
+  const jsonString = str.substring(str.indexOf('{'), str.lastIndexOf('}') + 1);
+
+  return JSON.parse(jsonString);
+}
