+++ packages/salesforcedx-utils/src/helpers/index.ts
@@ -1,0 +1,1 @@
+export { extractJsonObject } from './extractJsonObject';
