+++ packages/salesforcedx-utils/src/index.ts
@@ -1,6 +1,7 @@
+export * from './';
 export * from './cli';
 export * from './constants';
+export * from './helpers';
 export * from './i18n';
-export * from './';
 export * from './requestService';
 export * from './types';
